refactor: move home page to a separate component, add layout, add routing
