+++ src/components/Header/Header.tsx
@@ -1,28 +1,50 @@
+import { PaintBrushIcon } from '@heroicons/react/24/outline';
+import { NavLink, useNavigate } from 'react-router-dom';
+
+import { ROUTES } from '../../constants/routes.ts';
 import { Button } from '../Button/Button.tsx';
 import { Input } from '../Input/Input.tsx';
 
 type Props = {
-  onChange: (value: string) => void;
-  onSearch: () => void;
-  value: string;
+  onChange?: (value: string) => void;
+  onSearch?: () => void;
+  value?: string;
 };
 
 export const Header = ({ onChange, onSearch, value }: Props) => {
   const handleSubmit = (e: React.FormEvent) => {
-    e.preventDefault(); // Не даём странице перегружаться
-    onSearch();
+    e.preventDefault();
+    if (onSearch) onSearch();
   };
+  const showForm = onChange && value && onSearch;
+
+  const navigate = useNavigate();
 
   return (
     <header className={'bg-stone-100 p-6 shadow-md w-screen'}>
-      <div className="max-w-7xl mx-auto px-4">
-        <form
-          className={'flex justify-center items-center gap-3.5'}
-          onSubmit={handleSubmit}
-        >
-          <Input onChangeHandler={onChange} value={value} />
-          <Button title={'Search'} type={'submit'} />
-        </form>
+      <div className="max-w-6xl mx-auto px-4">
+        <div className={'flex items-center justify-between'}>
+          <PaintBrushIcon
+            className={'hover:text-lime-600 cursor-pointer'}
+            height={30}
+            onClick={() => navigate(ROUTES.HOME)}
+            width={30}
+          />
+          <NavLink to={ROUTES.ABOUT}>
+            <span className={'hover:text-lime-600 transition-colors text-lg'}>
+              ABOUT
+            </span>
+          </NavLink>
+          {showForm && (
+            <form
+              className={'flex justify-center items-center gap-3.5'}
+              onSubmit={handleSubmit}
+            >
+              <Input onChangeHandler={onChange} value={value} />
+              <Button title={'Search'} type={'submit'} />
+            </form>
+          )}
+        </div>
       </div>
     </header>
   );
