+++ src/constants/routes.ts
@@ -1,3 +1,4 @@
 export const ROUTES = {
+  ABOUT: '/about',
   HOME: '/',
 };
