+++ src/index.tsx
@@ -2,9 +2,9 @@ import { StrictMode } from 'react';
 import { createRoot } from 'react-dom/client';
 
 import './index.css';
-import { BrowserRouter, Route, Routes } from 'react-router-dom';
+import { BrowserRouter } from 'react-router-dom';
 
-import { App } from './App.tsx';
+import App from './App.tsx';
 import { ErrorBoundary } from './components/ErrorBoundary/ErrorBoundary.tsx';
 import { Fallback } from './components/Fallback/Fallback.tsx';
 
@@ -18,11 +18,7 @@ createRoot(rootElement).render(
   <StrictMode>
     <BrowserRouter>
       <ErrorBoundary fallback={<Fallback />}>
-        <Routes>
-          <Route element={<App />} path="/" />
-          <Route element={<App />} path=":page" />
-          <Route element={<App />} path=":page/:detailsId" />
-        </Routes>
+        <App />
       </ErrorBoundary>
     </BrowserRouter>
   </StrictMode>
